evalai-python/pyproject.toml
[project]
name = "evalai-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.2",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "paddleocr>=3.4.0",
    "paddlepaddle>=3.3.0",
    "pillow>=12.1.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "redis>=7.3.0",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.41.0",
]
